fix: replace unsafe numeric casts with checked conversions (#791)
Fixes #679, fixes #689.

- weather.rs: last_check_hour field/accessor changed from Option<u32> to
  Option<i64>; tick() now computes current_hour via now.timestamp()/3600
  without the as-u32 cast. For game dates in 1820 the timestamp is a large
  negative i64 — the previous as u32 silently wrapped, producing a wrong
  hour value and breaking the once-per-hour dedup guard.
- debug_snapshot.rs: WeatherEngineSnapshot.last_check_hour updated to match.
- pipeline.rs: three usize→u32 casts guarded with u32::try_from().expect().
- connections.rs: debug-log best_dist cast changed from as u32 to as u64
  (distances are always positive; u64 avoids truncation on large values).

+++ crates/parish-core/src/debug_snapshot.rs
@@ -123,7 +123,7 @@ pub struct WeatherDebug {
     /// Minimum duration before a transition is allowed (game-hours).
     pub min_duration_hours: f64,
     /// Game-hour cursor of the last transition evaluation, if any.
-    pub last_check_hour: Option<u32>,
+    pub last_check_hour: Option<i64>,
 }
 
 /// World graph summary for debug display.
